# docs/swagger.yaml
components:
  schemas:
    array-assessment_internal_dto.CreateAccountRequest:
      properties:
        account_holder_name:
          maxLength: 100
          minLength: 1
          type: string
        account_number:
          type: string
        account_type:
          enum:
          - CHECKING
          - SAVINGS
          - MONEY_MARKET
          type: string
        routing_number:
          type: string
      required:
      - account_holder_name
      - account_number
      - account_type
      - routing_number
      type: object
...
    array-assessment_internal_errors.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/array-assessment_internal_errors.ErrorDetail'
      type: object
    array-assessment_internal_models.Account:
      properties:
        account_number:
          type: string
        account_type:
          type: string
        balance:
          type: number
        closed_at:
          type: string
        created_at:
          type: string
        currency:
          type: string
        deleted_at:
          $ref: '#/components/schemas/gorm.DeletedAt'
        id:
          type: string
        interest_rate:
          type: number
        routing_number:
          type: string
        status:
          type: string
        updated_at:
          type: string
        user_id:
          type: string
      type: object
...
    message:
      properties:
        message:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Production-quality banking REST API for developer assessment and interviewing.
    Provides core banking functionality including identity management, account operations,
    customer management, and transaction processing.
  title: Array Banking API
  version: "1.0"
openapi: 3.1.0
paths:
  /accounts:
    get:
      description: Retrieve all bank accounts belonging to the authenticated user
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/array-assessment_internal_models.Account'
                type: array
          description: List of user's accounts
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get all user accounts
      tags:
      - Accounts
    post:
      description: Create a new bank account (checking, savings, or money_market)
        with optional initial deposit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/array-assessment_internal_dto.CreateAccountRequest'
        description: Account creation details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_dto.CreateAccountResponse'
          description: Account created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: TRANSACTION_002 - Invalid initial deposit amount
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - Accounts
  /accounts/{accountId}:
    delete:
      description: Permanently close an account. Account must have zero balance to
        be closed.
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/message'
                properties:
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    type: object
                type: object
          description: Account closed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: VALIDATION_003 - Invalid account ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: ACCOUNT_005 - Account has non-zero balance
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Close account
      tags:
      - Accounts
    get:
      description: Retrieve detailed information about a specific account belonging
        to the authenticated user
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_models.Account'
          description: Account details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: VALIDATION_003 - Invalid account ID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - Accounts
  /accounts/{accountId}/status:
    patch:
      description: Update the status of an account (active, inactive, frozen, closed)
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/array-assessment_internal_dto.UpdateAccountStatusRequest'
        description: New account status
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_models.Account'
          description: Updated account details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid request body or account ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Update account status
      tags:
      - Accounts
...
  /admin/users/{userId}/accounts:
    get:
      description: Admin endpoint to retrieve all accounts for a specific user
      parameters:
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/array-assessment_internal_models.Account'
                type: array
          description: List of user's accounts
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: VALIDATION_003 - Invalid user ID format
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: AUTH_005 - Requires admin role
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/array-assessment_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get user accounts by user ID (admin)
      tags:
      - Admin
...
servers:
- url: /api/v1